requirements.txt
biopython==1.79
ete3==3.1.2
pandas==1.4.3
pdb-tools==2.5.0
pyqt5==5.15.9
requests==2.28.1
scikit-learn==1.1.2
seaborn==0.11.2
streamlit==1.18.1
altair==4.2
